NLP basis - words
The first steps with natural language processing lie with understanding the basic units of text data – words. Because for any single document, words (in various forms) are combine to develop an idea or a concept. In this short workbook, I demonstrate how you can take any arbitrary piece of text and process it for further use.